requirements.txt
Django ~= 4.2.4
django-comments-dab ~= 2.8.0
django-star-ratings ~= 0.9.2
django-allauth ~= 0.55.2
Pillow ~= 10.0.0
djangorestframework ~= 3.14.0